Initial extraction of template for website
Right now, files using the template are headed with yaml frontmatter
in '---'

The template needs some logic to handle when to display the stylesheet
link.

I have taken the rightnavs and put them all into the template.  The hope
is that one can be randomly selected for display when the browser loads
the page.  This logic has not been implemented.