Working on advancements
It is going to be easier to use a JS lightbox library than to build
static pages for paginating images.  If thumbnails are needed we can
generate them during build.  (This is unlikely, as for most browsers,
the lightbox library should handle that nicely.  The only time that
we would really care is if we care about behavior on old browsers,
and it's possible that even there, a single page filled with full
size images might work well enough.)

The advancement generation code here is initial.  I haven't gotten
it to build the collections correctly, or adjusted the template
to render them in any reasonable way.  This is still going to need
some more work.  Because of this, advancements is still disabled in
build.boot.