Header images and file name templating.
Fixed the issue where the eclipse always displayed as the header
image.  Now, it is loaded from a template in header_images based
on the header-image frontmatter tag.  The default image has also
been changed to oxbow.

Also, while working on this, I have made the templating more
flexible.  It now looks up the file indicated by the :template
key, after appending ".html_template" to the end.  Additionally,
as a result of this change, the verification logic has been
updated to handle different expected keys per template.

Furthermore, it is now possible to set default values for yaml
keys in site/core.clj.  This might be handled better using the
global metadata feature in perun, but that is a change that can
be done later.